Replaced akamai library + handle events manually
