+++ example/ios/ma_sdk.framework/Headers/ma_sdk.h
@@ -1,0 +1,27 @@
+
+/*
+ * ma_sdk.h
+ * Version - 0.1
+ *
+ * This file is part of the Media Analytics, http://www.akamai.com
+ * Media Analytics is a proprietary Akamai software that you may use and modify per the license agreement here:
+ * http://www.akamai.com/product/licenses/mediaanalytics.html
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS “AS IS” AND ANY EXPRESS OR IMPLIED WARRANTIES,
+ * INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+ *
+ *
+ * Created by El Visy on 25th January 2016.
+ *
+ */
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for ma_sdk.
+FOUNDATION_EXPORT double ma_sdkVersionNumber;
+
+//! Project version string for ma_sdk.
+FOUNDATION_EXPORT const unsigned char ma_sdkVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <ma_sdk/PublicHeader.h>
+
+
